clang fixes

+++ src/drivers/7-segment.cpp
@@ -61,103 +61,103 @@ uint8_t SevenSegment::seg(char data, bool reverse)
         case 0x06 : return f;
         case 0x07 : return g;
 
-        case '0' : return a+b+c+d+e+f;
-        case '1' : return b+c;
-        case '2' : return a+b+g+e+d;
-        case '3' : return a+b+g+c+d;
-        case '4' : return f+b+g+c;
-        case '5' : return a+f+g+c+d;
-        case '6' : return a+f+g+e+c+d;
-        case '7' : return a+b+c;
-        case '8' : return a+f+b+g+e+c+d;
-        case '9' : return a+f+b+g+c+d;
+        case '0' : return uint8_t(a+b+c+d+e+f);
+        case '1' : return uint8_t(b+c);
+        case '2' : return uint8_t(a+b+g+e+d);
+        case '3' : return uint8_t(a+b+g+c+d);
+        case '4' : return uint8_t(f+b+g+c);
+        case '5' : return uint8_t(a+f+g+c+d);
+        case '6' : return uint8_t(a+f+g+e+c+d);
+        case '7' : return uint8_t(a+b+c);
+        case '8' : return uint8_t(a+f+b+g+e+c+d);
+        case '9' : return uint8_t(a+f+b+g+c+d);
 
         case ' ' : return 0;
         case '.' : return dp;
         case ',' : return dp;
-        case '+' : return b+g+c;
+        case '+' : return uint8_t(b+g+c);
         case '-' : return g;
         case '=' : return g+d;
         case '"' : return f+b;
         case '\'': return b;
-        case '!' : return b+c+dp;
+        case '!' : return uint8_t(b+c+dp);
         case '|' : return f+e;
         case '(' :
         case '{' :
-        case '[' : return a+f+e+d;
+        case '[' : return uint8_t(a+f+e+d);
         case ')' :
         case '}' :
-        case ']' : return a+b+c+d;
-        case '?' : return a+b+g+e;
-        case '*' : return f+b+g+e+c;
-        case '#' : return a+f+b+g;
-        case '/' : return b+g+e;
-        case '\\': return f+g+c;
-        case ':' : return g+d;
-        case ';' : return g+d;
-        case '$' : return a+f+g+c+d;
-        case '%' : return b+g+e;
-        case '&' : return a+f+b+g+e;
+        case ']' : return uint8_t(a+b+c+d);
+        case '?' : return uint8_t(a+b+g+e);
+        case '*' : return uint8_t(f+b+g+e+c);
+        case '#' : return uint8_t(a+f+b+g);
+        case '/' : return uint8_t(b+g+e);
+        case '\\': return uint8_t(f+g+c);
+        case ':' : return uint8_t(g+d);
+        case ';' : return uint8_t(g+d);
+        case '$' : return uint8_t(a+f+g+c+d);
+        case '%' : return uint8_t(b+g+e);
+        case '&' : return uint8_t(a+f+b+g+e);
         case '_' : return d;
-        case '^' : return a+f+b;
+        case '^' : return uint8_t(a+f+b);
         case '`' : return f;
-        case '@' : return a+f+b+g+e+d;
-        case '>' : return g+c+d;
-        case '<' : return g+e+d;
+        case '@' : return uint8_t(a+f+b+g+e+d);
+        case '>' : return uint8_t(g+c+d);
+        case '<' : return uint8_t(g+e+d);
         case '~' : return a;
 
         case 'a' :
-        case 'A' : return a+f+b+g+e+c;
-        case 'b' : return f+g+e+c+d;
-        case 'B' : return f+g+e+c+d;
-        case 'c' : return g+e+d;
-        case 'C' : return a+f+e+d;
-        case 'd' : return b+g+e+c+d;
-        case 'D' : return a+f+b+e+c+d;
+        case 'A' : return uint8_t(a+f+b+g+e+c);
+        case 'b' : return uint8_t(f+g+e+c+d);
+        case 'B' : return uint8_t(f+g+e+c+d);
+        case 'c' : return uint8_t(g+e+d);
+        case 'C' : return uint8_t(a+f+e+d);
+        case 'd' : return uint8_t(b+g+e+c+d);
+        case 'D' : return uint8_t(a+f+b+e+c+d);
         case 'e' :
-        case 'E' : return a+f+g+e+d;
+        case 'E' : return uint8_t(a+f+g+e+d);
         case 'f' :
-        case 'F' : return a+f+g+e;
+        case 'F' : return uint8_t(a+f+g+e);
         case 'g' :
-        case 'G' : return a+f+b+g+c+d;
-        case 'h' : return f+g+e+c;
-        case 'H' : return f+b+g+e+c;
+        case 'G' : return uint8_t(a+f+b+g+c+d);
+        case 'h' : return uint8_t(f+g+e+c);
+        case 'H' : return uint8_t(f+b+g+e+c);
         case 'i' : 
         case 'I' : return b+c;
         case 'j' :
-        case 'J' : return b+c+d;
-        case 'k' : return a+f+g+e+c;
-        case 'K' : return f+b+g+e+c;
+        case 'J' : return uint8_t(b+c+d);
+        case 'k' : return uint8_t(a+f+g+e+c);
+        case 'K' : return uint8_t(f+b+g+e+c);
         case 'l' :
-        case 'L' : return f+e+d;
-        case 'm' : return g+e+c;
-        case 'M' : return a+f+b+e+c;
-        case 'n' : return g+e+c;
-        case 'N' : return a+f+b+e+c;
-        case 'o' : return g+e+c+d;
-        case 'O' : return a+f+b+e+c+d;
+        case 'L' : return uint8_t(f+e+d);
+        case 'm' : return uint8_t(g+e+c);
+        case 'M' : return uint8_t(a+f+b+e+c);
+        case 'n' : return uint8_t(g+e+c);
+        case 'N' : return uint8_t(a+f+b+e+c);
+        case 'o' : return uint8_t(g+e+c+d);
+        case 'O' : return uint8_t(a+f+b+e+c+d);
         case 'p' :
-        case 'P' : return a+f+b+g+e;
+        case 'P' : return uint8_t(a+f+b+g+e);
         case 'q' :
-        case 'Q' : return a+f+b+g+c;
+        case 'Q' : return uint8_t(a+f+b+g+c);
         case 'r' :
         case 'R' : return g+e;
         case 's' :
-        case 'S' : return a+f+g+c+d;
+        case 'S' : return uint8_t(a+f+g+c+d);
         case 't' : 
-        case 'T' : return f+g+e+d;
-        case 'u' : return e+c+d;
-        case 'U' : return f+b+e+c+d;
-        case 'v' : return e+c+d;
-        case 'V' : return f+b+e+c+d;
-        case 'w' : return e+c+d;
-        case 'W' : return f+b+e+c+d;
+        case 'T' : return uint8_t(f+g+e+d);
+        case 'u' : return uint8_t(e+c+d);
+        case 'U' : return uint8_t(f+b+e+c+d);
+        case 'v' : return uint8_t(e+c+d);
+        case 'V' : return uint8_t(f+b+e+c+d);
+        case 'w' : return uint8_t(e+c+d);
+        case 'W' : return uint8_t(f+b+e+c+d);
         case 'x' :
-        case 'X' : return f+b+g+e+c;
+        case 'X' : return uint8_t(f+b+g+e+c);
         case 'y' :
-        case 'Y' : return f+b+g+c+d;
+        case 'Y' : return uint8_t(f+b+g+c+d);
         case 'z' :
-        case 'Z' : return a+b+g+e+d;
+        case 'Z' : return uint8_t(a+b+g+e+d);
 
         default : return 0;
     }
